Bug fix for `automatic_serialization` in Zend
This bug fix addresses an issue reported by @kVasiliev involving
`Zend_Cache_Core` with `automatic_serialization = true` which no unit
tests existed for Backend testing with this setting. Once the setting
was enabled it was discovered the
`Rediska_Command_GetFromHash::parseResponse` unserializes all responses.
To control this behavior a property and public method was added to the
`Rediska_Serializer`. To counteract the default behaviour applying
advice from @shumkov the Backend switches the serializer adapter to
`toString` and once it acquires the data from the hash it reinstates the
previos adapter to the serializer.

Additional `catch` calls were also added to transaction handling for
`Rediska_Transaction_Aborted_Exception` for all `execute` calls as well
as additional seperation of transaction usage.