Added the awareness of empty transactions.
I moved the preamble for the `Rediska_Transaction::__toString` call for
both populated and empty transaction to a class constant. This was to
ensure uniformity of a potential future change no overlooking or
breaking the `Rediska_Zend_Cache_Backend_Redis` classes use in the
future. Two methods added that ensure transactions are executed on the
proper connection by key for multiple connection uses. When executing
the transaction if a transaction is empty `Rediska_Transaction::discard`
is called. This allows for more awareness of the
`Rediska_Transaction_AbortedException` and prevents it from being
erroneously thrown in execution is made with no commands being run.